@@ -12,66 +12,68 @@
     "start": "npm run build && node server/index.js",
     "build": "npm run build:setup && npm run build:client",
     "build:setup": "rm -rf public/*",
-    "build:client": "webpack --config webpack.config.js --env.production"
+    "build:client": "webpack --config webpack.config.js --env production"
   },
   "author": "Miguel Peixe <miguel@peixe.co>",
   "homepage": "https://github.com/miguelpeixe/jitsi-provisioner",
   "license": "Apache-2.0",
   "dependencies": {
-    "@feathersjs/authentication": "^4.5.7",
-    "@feathersjs/authentication-client": "^4.5.7",
-    "@feathersjs/authentication-local": "^4.5.7",
-    "@feathersjs/express": "^4.5.7",
-    "@feathersjs/feathers": "^4.5.7",
-    "@feathersjs/rest-client": "^4.5.7",
-    "@feathersjs/socketio": "^4.5.7",
-    "@feathersjs/socketio-client": "^4.5.7",
+    "@feathersjs/authentication": "^4.5.9",
+    "@feathersjs/authentication-client": "^4.5.9",
+    "@feathersjs/authentication-local": "^4.5.9",
+    "@feathersjs/express": "^4.5.9",
+    "@feathersjs/feathers": "^4.5.9",
+    "@feathersjs/rest-client": "^4.5.9",
+    "@feathersjs/socketio": "^4.5.9",
+    "@feathersjs/socketio-client": "^4.5.9",
     "@hapi/joi": "^17.1.1",
-    "@jitsi-provisioner/aws-utils": "0.0.2",
-    "archiver": "^5.0.0",
-    "axios": "^0.19.2",
+    "@jitsi-provisioner/api": "0.0.1",
+    "archiver": "^5.0.2",
+    "axios": "^0.20.0",
     "cloudflare": "^2.7.0",
-    "feathers-hooks-common": "^5.0.3",
+    "feathers-hooks-common": "^5.0.4",
     "feathers-nedb": "^5.1.0",
     "jsonwebtoken": "^8.5.1",
-    "lodash": "^4.17.19",
-    "moment": "^2.27.0",
-    "nanoid": "^3.1.11",
+    "lodash": "^4.17.20",
+    "moment": "^2.29.1",
+    "nanoid": "^3.1.12",
     "nedb": "^1.8.0",
-    "parse-domain": "^3.0.2",
-    "preact": "^10.4.6",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "parse-domain": "^3.0.3",
+    "preact": "^10.5.5",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "rsuite": "^4.8.0",
-    "socket.io-client": "^2.3.0",
-    "styled-components": "^5.1.1",
+    "rsuite": "^4.8.4",
+    "socket.io-client": "^2.3.1",
+    "styled-components": "^5.2.0",
     "winston": "^3.3.3"
   },
   "devDependencies": {
-    "@babel/cli": "^7.10.5",
-    "@babel/core": "^7.10.5",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-export-default-from": "^7.10.4",
-    "@babel/plugin-transform-react-jsx": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.10.5",
-    "@babel/plugin-transform-spread": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
+    "@babel/cli": "^7.12.1",
+    "@babel/core": "^7.12.3",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-export-default-from": "^7.12.1",
+    "@babel/plugin-transform-react-jsx": "^7.12.1",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/plugin-transform-spread": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/preset-react": "^7.12.1",
     "babel-loader": "^8.1.0",
     "connect-history-api-fallback": "^1.6.0",
-    "css-loader": "^4.0.0",
-    "file-loader": "^6.0.0",
-    "html-webpack-plugin": "^4.3.0",
-    "less-loader": "^6.2.0",
-    "mini-css-extract-plugin": "^0.9.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "style-loader": "^1.2.1",
-    "terser-webpack-plugin": "^3.0.8",
-    "webpack": "^4.44.0",
-    "webpack-bundle-analyzer": "^3.8.0",
-    "webpack-cli": "^3.3.12",
+    "css-loader": "^5.0.0",
+    "css-minimizer-webpack-plugin": "^1.1.5",
+    "file-loader": "^6.1.1",
+    "html-webpack-plugin": "^4.5.0",
+    "less": "^3.12.2",
+    "less-loader": "^7.0.2",
+    "mini-css-extract-plugin": "^1.1.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.4",
+    "style-loader": "^2.0.0",
+    "terser-webpack-plugin": "^5.0.0",
+    "webpack": "^5.1.3",
+    "webpack-bundle-analyzer": "^3.9.0",
+    "webpack-cli": "^4.1.0",
     "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0"
   },
